requirements.txt
Django==4.2.7
django-mssql-backend==2.8.1
pyodbc==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0